Always write chapter string to .vcp
When "Create chapters per clip" is switched off, VidCutter writes `.vcp` 
files with _no_ fourth argument, which it is then unable to read 
(reporting "Invalid line" for every line in the file after the first). 
This change causes the fourth argument of the project file to be written 
as `""` whenever the create-chapters preference is turned off, ensuring 
that the resulting `.vcp` file can be read by VidCutter.